feat: decouple login and register page
- split into two pages to facilitate users to share URLs
- remove password autocomplete poten potentially
